Initial parsing of request parameters is case-insensitive
There is a security issue where accessing 'ADMIN' vs 'admin' would result in complete access to the application. This might be mitigated with routing or a beforeFilter(), but it is best to not depend on the application taking care of this.

In the future, we may have a case-insensitive version, or allow for checking of case-sensitivity only on certain request parameter keys.